move layout from component to src directory
